+++ source/extensions/omni.isaac.core/omni/isaac/core/materials/__init__.py
@@ -10,3 +10,4 @@ from omni.isaac.core.materials.visual_material import VisualMaterial
 from omni.isaac.core.materials.preview_surface import PreviewSurface
 from omni.isaac.core.materials.physics_material import PhysicsMaterial
 from omni.isaac.core.materials.omni_glass import OmniGlass
+from omni.isaac.core.materials.omni_pbr import OmniPBR