+++ source/extensions/omni.isaac.core/omni/isaac/core/objects/capsule.py
@@ -12,6 +12,7 @@ from omni.isaac.core.prims.rigid_prim import RigidPrim
 from omni.isaac.core.prims.geometry_prim import GeometryPrim
 from omni.isaac.core.materials import PreviewSurface
 from omni.isaac.core.materials import PhysicsMaterial
+from omni.isaac.core.utils.string import find_unique_string_name
 from pxr import UsdGeom, Gf
 from omni.isaac.core.utils.prims import get_prim_at_path, is_prim_path_valid
 from omni.isaac.core.utils.stage import get_current_stage
@@ -61,7 +62,10 @@ class VisualCapsule(GeometryPrim):
             if visual_material is None:
                 if color is None:
                     color = np.array([0.5, 0.5, 0.5])
-                visual_material = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
+                visual_prim_path = find_unique_string_name(
+                    intitial_name="/World/Looks/visual_material", is_unique_fn=lambda x: not is_prim_path_valid(x)
+                )
+                visual_material = PreviewSurface(prim_path=visual_prim_path, color=color)
             VisualCapsule.apply_visual_material(self, visual_material)
         return
 
@@ -176,12 +180,19 @@ class DynamicCapsule(RigidPrim, GeometryPrim):
             if visual_material is None:
                 if color is None:
                     color = np.array([0.5, 0.5, 0.5])
-                visual_material = PreviewSurface(prim_path=prim_path + "/visual_material", color=color)
+                visual_prim_path = find_unique_string_name(
+                    intitial_name="/World/Looks/visual_material", is_unique_fn=lambda x: not is_prim_path_valid(x)
+                )
+                visual_material = PreviewSurface(prim_path=visual_prim_path, color=color)
             DynamicCapsule.apply_visual_material(self, visual_material)
 
         if physics_material_path is None:
+            physics_material_path = find_unique_string_name(
+                intitial_name="/World/Physics_Materials/physics_material",
+                is_unique_fn=lambda x: not is_prim_path_valid(x),
+            )
             my_physics_material = PhysicsMaterial(
-                prim_path=prim_path + "/physics_material",
+                prim_path=physics_material_path,
                 dynamic_friction=dynamic_friction,
                 static_friction=static_friction,
                 restitution=restitution,